Tweaking based on feedback

+++ 03-ingress-controller/certificates/add-certificates.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if [ ! -f "env.sh" ]; then
+    echo "ERROR: File env.sh not found."
+    exit 1
+fi
+
+source ./env.sh
+
 for cert in portal api notes; do
     if kubectl get secret ${cert}-tls &> /dev/null; then
         kubectl delete secret ${cert}-tls
